+++ OneBlink.SDK/OneBlinkHttpClient.cs
@@ -3,6 +3,7 @@ using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
+using System.IO;
 
 namespace OneBlink.SDK
 {
@@ -62,6 +63,26 @@ namespace OneBlink.SDK
           throw new OneBlinkAPIException(apiErrorResponse);
         }
         return httpResponseMessage.Content;
+      }
+    }
+
+    public async Task<Stream> StreamRequest(HttpRequestMessage httpRequestMessage, string userToken = null)
+    {
+      // generate token
+      string token = !String.IsNullOrEmpty(userToken) ? userToken : Token.GenerateJSONWebToken(accessKey, secretKey, expiryInSeconds);
+      using (HttpClient httpClient = new HttpClient())
+      {
+        httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
+        httpClient.DefaultRequestHeaders.Add("User-Agent", ".Net OneBlink.SDK / " + GetType().Assembly.GetName().Version.ToString());
+
+        HttpResponseMessage httpResponseMessage = await httpClient.SendAsync(httpRequestMessage);
+        if (!httpResponseMessage.IsSuccessStatusCode)
+        {
+          string result = await httpResponseMessage.Content.ReadAsStringAsync();
+          APIErrorResponse apiErrorResponse = JsonConvert.DeserializeObject<APIErrorResponse>(result);
+          throw new OneBlinkAPIException(apiErrorResponse);
+        }
+        return await httpResponseMessage.Content.ReadAsStreamAsync();
       }
     }
   }