+++ OneBlink.SDK/models/FormSubmissionEventConfiguration.cs
@@ -101,5 +101,9 @@ namespace OneBlink.SDK.Model
         {
             get; set;
         }
+        public FormSubmissionEventTrimUriOption author
+        {
+            get; set;
+        }
     }
 }