+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - `displayAsCurrency` to `FormElement`
 - `storageType` to `FormElement`
 - `formsAppEnvironment` to `FormsClient Search`
+- `GetFormSubmissionAttachment` to `FormsClient`
 
 ## 2.1.0 (2021-04-27)