+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - `formsAppEnvironment` to `FormsClient Search`
 - `GetFormSubmissionAttachment` to `FormsClient`
 - `CreateSubmissionAttachment` to `FormsClient`
+- `author` to `FormSubmissionEventConfiguration`
 
 ## 2.1.0 (2021-04-27)