+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - `GetFormSubmissionAttachment` to `FormsClient`
 - `CreateSubmissionAttachment` to `FormsClient`
 - `author` to `FormSubmissionEventConfiguration`
+- `recaptchaIntegrationDomainId` to `FormsAppBase`
 
 ## 2.1.0 (2021-04-27)