+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 2.2.0 (2021-05-13)
+
 ### Fixed
 
 - `ConditionallyShowPredicate` class missing BETWEEN `type`